% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agm.R
\name{agm}
\alias{agm}
\title{Arithmetic-geometric mean}
\usage{
agm(x, y)
}
\arguments{
\item{x, y}{complex numbers}
}
\value{
A complex number, the arithmetic-geometric mean of \code{x} 
  and \code{y}.
}
\description{
Evaluation of the arithmetic-geometric mean of two 
  complex numbers.
}
\examples{
agm(1, sqrt(2))
2*pi^(3/2)*sqrt(2) / gamma(1/4)^2
}
