% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am.R
\name{am}
\alias{am}
\title{Amplitude function}
\usage{
am(u, k)
}
\arguments{
\item{u}{complex number}

\item{k}{elliptic modulus, a complex number}
}
\value{
A complex number.
}
\description{
Evaluation of the amplitude function.
}
\examples{
library(Carlson)
phi <- 1 + 1i
k <- 2
u <- elliptic_F(phi, k^2)
am(u, k) # should be phi
}
