% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipticAlpha.R
\name{ellipticAlpha}
\alias{ellipticAlpha}
\title{Elliptic alpha function}
\usage{
ellipticAlpha(z)
}
\arguments{
\item{z}{a complex number}
}
\value{
A complex number.
}
\description{
Evaluates the elliptic alpha function.
}
\references{
Weisstein, Eric W. 
\href{https://mathworld.wolfram.com/EllipticAlphaFunction.html}{"Elliptic Alpha Function"}.
}
