% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weierstrassStuff.R
\name{ellipticInvariants}
\alias{ellipticInvariants}
\title{Elliptic invariants}
\usage{
ellipticInvariants(omega1omega2)
}
\arguments{
\item{omega1omega2}{the half-periods, a vector of two complex numbers}
}
\value{
The elliptic invariants, a vector of two complex numbers.
}
\description{
Elliptic invariants from half-periods.
}
