% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eisenstein.R
\name{En}
\alias{En}
\title{Eisenstein series}
\usage{
En(n, q)
}
\arguments{
\item{n}{the weight, can be 2, 4 or 6}

\item{q}{nome, complex number with modulus smaller than one, but not 
a negative real number}
}
\value{
A complex number, the value of the Eisenstein series.
}
\description{
Evaluation of Eisenstein series with weight 2, 4 or 6.
}
