% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta-dedekind.R
\name{eta}
\alias{eta}
\title{Dedekind eta function}
\usage{
eta(tau)
}
\arguments{
\item{tau}{a complex number with strictly positive imaginary part}
}
\value{
A complex number.
}
\description{
Evaluation of the Dedekind function.
}
\examples{
eta(2i)
gamma(1/4) / 2^(11/8) / pi^(3/4)
}
