% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weierstrassStuff.R
\name{halfPeriods}
\alias{halfPeriods}
\title{Half-periods}
\usage{
halfPeriods(g2g3)
}
\arguments{
\item{g2g3}{the elliptic invariants, a vector of two complex numbers}
}
\value{
The half-periods, a vector of two complex numbers.
}
\description{
Half-periods from elliptic invariants.
}
