% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_script.R
\name{jacobi}
\alias{jacobi}
\title{methode de jacobi}
\usage{
jacobi(matrice, xo, B, nomb_ite)
}
\arguments{
\item{matrice}{une matrice carree}

\item{xo}{approximation initiale}

\item{B}{second membre du systeme}

\item{nomb_ite}{nombres d iterations a effectuer}
}
\description{
La methode de jacobi, due au mathematicien allemand Karl Jacobi, est une methode iterative
de resolution d un systeme matriciel de la forme Ax=B. Pour cela, on utilise une suite x^k
qui converge vers un point fixe x, solution du systeme d equations lineaires.
On cherche a construire l algorithme pour x^0 donne, la suite x^(k+1)=F(x^k) avec k appartenant N.
}
\examples{
jacobi(matrix(c(9,1,1,4,6,-2,1,0,-6),3,3),
matrix(c(0,0,0),3,1),matrix(c(-17,4,14),3,1),2)
}
\author{
DJEUMENI Alexia
}
