% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtheta.R
\name{jtheta1}
\alias{jtheta1}
\alias{ljtheta1}
\title{Jacobi theta function one}
\usage{
jtheta1(z, tau = NULL, q = NULL)

ljtheta1(z, tau = NULL, q = NULL)
}
\arguments{
\item{z}{complex number}

\item{tau}{lattice parameter, a complex number with strictly positive 
imaginary part; the two complex numbers \code{tau} and \code{q} are 
related by \code{q = exp(1i*pi*tau)}, and only one of them must be 
supplied}

\item{q}{the nome, a complex number whose modulus is strictly less than one, 
and which is not zero nor a negative real number}
}
\value{
A complex number; \code{jtheta1} evaluates the first Jacobi theta 
  function and \code{ljtheta1} evaluates its logarithm.
}
\description{
Evaluates the first Jacobi theta function.
}
\examples{
jtheta1(1 + 1i, q = exp(-pi/2))
}
