% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jtheta.R
\name{jtheta3}
\alias{jtheta3}
\alias{ljtheta3}
\title{Jacobi theta function three}
\usage{
jtheta3(z, tau = NULL, q = NULL)

ljtheta3(z, tau = NULL, q = NULL)
}
\arguments{
\item{z}{complex number}

\item{tau}{lattice parameter, a complex number with strictly positive 
imaginary part; the two complex numbers \code{tau} and \code{q} are 
related by \code{q = exp(1i*pi*tau)}, and only one of them must be 
supplied}

\item{q}{the nome, a complex number whose modulus is strictly less than one, 
and which is not zero nor a negative real number}
}
\value{
A complex number; \code{jtheta3} evaluates the third Jacobi theta 
  function and \code{ljtheta3} evaluates its logarithm.
}
\description{
Evaluates the third Jacobi theta function.
}
\examples{
jtheta3(1 + 1i, q = exp(-pi/2))
}
