% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein-j.R
\name{kleinj}
\alias{kleinj}
\alias{kleinjinv}
\title{Klein j-function and its inverse}
\usage{
kleinj(tau)

kleinjinv(j)
}
\arguments{
\item{tau}{a complex number with strictly positive imaginary part}

\item{j}{a complex number}
}
\value{
A complex number.
}
\description{
Evaluation of the Klein j-invariant function and its inverse.
}
\examples{
( j <- kleinj(2i) )
66^3
kleinjinv(j)
}
