% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta-neville.R
\name{theta.s}
\alias{theta.s}
\alias{theta.c}
\alias{theta.n}
\alias{theta.d}
\title{Neville theta functions}
\usage{
theta.s(z, tau = NULL, m = NULL)

theta.c(z, tau = NULL, m = NULL)

theta.n(z, tau = NULL, m = NULL)

theta.d(z, tau = NULL, m = NULL)
}
\arguments{
\item{z}{a complex number, vector, or matrix}

\item{tau}{complex number with strictly positive imaginary part; it is 
related to \code{m} and only one of them must be supplied}

\item{m}{the "parameter", square of the elliptic modulus; it is related to 
\code{tau} and only one of them must be supplied}
}
\value{
A complex number, vector or matrix.
}
\description{
Evaluation of the Neville theta functions.
}
