% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nome.R
\name{nome}
\alias{nome}
\title{Nome}
\usage{
nome(m)
}
\arguments{
\item{m}{the parameter, square of elliptic modulus}
}
\value{
A complex number.
}
\description{
The nome in function of the parameter \eqn{m}.
}
\examples{
nome(-2)
}
