% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-weierstrass.R
\name{pweierstrass}
\alias{pweierstrass}
\title{x}
\usage{
pweierstrass(z, g = NULL, omega = NULL, derivative = 0L)
}
\arguments{
\item{z}{x}

\item{g}{elliptic invariants}

\item{omega}{half-periods}

\item{derivative}{x}
}
\value{
x
}
\description{
x
}
\examples{
omega1 <- 1.4 - 1i
omega2 <- 1.6 + 0.5i
omega <- c(omega1, omega2)
e1 <- pweierstrass(omega1, omega = omega)
e2 <- pweierstrass(omega2, omega = omega)
e3 <- pweierstrass(-omega1-omega2, omega = omega)
e1 + e2 + e3 # should be 0
}
