% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RogersRamanajuan.R
\name{RRa}
\alias{RRa}
\title{Alternating Rogers-Ramanajuan continued fraction}
\usage{
RRa(q)
}
\arguments{
\item{q}{the nome, a complex number whose modulus is strictly less than one, 
and which is not zero nor a negative real number}
}
\value{
A complex number
}
\description{
Evaluates the alternating Rogers-Ramanajuan continued fraction.
}
\note{
This function is sometimes denoted by \eqn{S}.
}
