% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeta.R
\name{zetaw}
\alias{zetaw}
\title{Weierstrass zeta function}
\usage{
zetaw(z, g = NULL, omega = NULL, tau = NULL)
}
\arguments{
\item{z}{complex number}

\item{g}{the elliptic invariants, a vector of two complex numbers; only 
one of \code{g}, \code{omega} and \code{tau} must be given}

\item{omega}{the half-periods, a vector of two complex numbers; only 
one of \code{g}, \code{omega} and \code{tau} must be given}

\item{tau}{the half-periods ratio; supplying \code{tau} is equivalent to 
supply \code{omega = c(1/2, tau/2)}}
}
\value{
A complex number.
}
\description{
Evaluation of the Weierstrass zeta function.
}
\examples{
# Mirror symmetry property:
z <- 1 + 1i
g <- c(1i, 1+2i)
zetaw(Conj(z), Conj(g))
Conj(zetaw(z, g))
}
