% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein-j.R
\name{kleinj}
\alias{kleinj}
\title{Klein j-function}
\usage{
kleinj(tau)
}
\arguments{
\item{tau}{a complex number with strictly positive imaginary part}
}
\value{
A complex number.
}
\description{
Evaluation of the Klein j-invariant function.
}
\examples{
kleinj(2i)
66^3
}
