% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{lambda}
\alias{lambda}
\title{Lambda modular function}
\usage{
lambda(tau)
}
\arguments{
\item{tau}{a complex number with strictly positive imaginary part}
}
\value{
A complex number.
}
\description{
Evaluation of the lambda modular function.
}
\note{
The lambda function is the square of the elliptic modulus.
}
\examples{
eta(2i)
gamma(1/4) / 2^(11/8) / pi^(3/4)
}
