% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{lambda}
\alias{lambda}
\title{Lambda modular function}
\usage{
lambda(tau)
}
\arguments{
\item{tau}{a complex number with strictly positive imaginary part, or 
a vector or matrix of such complex numbers}
}
\value{
A complex number, vector or matrix.
}
\description{
Evaluation of the lambda modular function.
}
\note{
The lambda function is the square of the elliptic modulus.
}
\examples{
x <- 2
lambda(1i*sqrt(x)) + lambda(1i*sqrt(1/x)) # should be one
}
