% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemniscate.R
\name{lemniscate}
\alias{lemniscate}
\alias{sl}
\alias{cl}
\alias{asl}
\alias{acl}
\alias{slh}
\alias{clh}
\title{Lemniscate functions}
\usage{
sl(z)

cl(z)

asl(z)

acl(z)

slh(z)

clh(z)
}
\arguments{
\item{z}{a real number or a complex number}
}
\value{
A complex number.
}
\description{
Lemniscate sine, cosine, arcsine, arccosine, hyperbolic sine, 
  and hyperbolic cosine functions.
}
\examples{
sl(1+1i) * cl(1+1i) # should be 1
}
