% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RogersRamanujan.R
\name{RR}
\alias{RR}
\title{Rogers-Ramanujan continued fraction}
\usage{
RR(q)
}
\arguments{
\item{q}{the nome, a complex number whose modulus is strictly less than one, 
and which is not zero nor a negative real number}
}
\value{
A complex number
}
\description{
Evaluates the Rogers-Ramanujan continued fraction.
}
\note{
This function is sometimes denoted by \eqn{R}.
}
