% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeta.R
\name{zetaw}
\alias{zetaw}
\title{Weierstrass zeta function}
\usage{
zetaw(z, g = NULL, omega = NULL, fix = FALSE)
}
\arguments{
\item{z}{complex number}

\item{g}{the elliptic invariants, a vector of two complex numbers; they are 
related to the half-periods (\code{omega}) and only one of \code{g} 
and \code{omega} must be given}

\item{omega}{the half-periods, a vector of two complex numbers; they are 
related to the elliptic invariants (\code{g}) and only one of \code{g} 
and \code{omega} must be given}

\item{fix}{Boolean; if \code{TRUE} and if there is \code{NaN} or \code{Inf} 
in the result, the function tries to get the correct result by applying 
some transformations}
}
\value{
A complex number.
}
\description{
Evaluation of the Weierstrass zeta function.
}
\examples{
# Mirror symmetry property:
z <- 1 + 1i
g <- c(1i, 1+2i)
zetaw(Conj(z), Conj(g))
Conj(zetaw(z, g))
}
